+++ Solutions/PowerShell.Commands/changelog.md
@@ -1,5 +1,8 @@
 # OfficeDevPnP.PowerShell Changelog #
 
+**2014-09-19**
+* Removed Url Parameters on Add-SPOFile and made Folder parameter mandatory.
+
 **2014-09-06**
 * Added new Set-SPOWeb cmdlet to set Title, SiteLogo, or AlternateCssUrl